Refactor tokenizer mode handling: add tokenizer_mode.h for f-string and t-string state management
